Fw_base - Disable Quick Settings on secure lockscreen (1/2)
This commit gets the following ones uptodate with marshmallow.
Adapted @KreAch3R

- Moved the necessary bits into their new functions
- isSecure() now explicity needs a userhandle (followed google's way
because USER_CURRENT didn't work)
- Implemented a mKeyguardOrShadeShowing boolean to stop the bug of
notification drawer (which reappeared in marshmallow even after the fix
because keyguardShowing changed.

Frameworks: Disable access to quick settings on secure lockscreen (1/2)

We have a lot security guys as user due of using the phones as business
phones...that
said a lot complains go into the direction of google how it is possible
to have
access to quick settings on a secure lockscreen which is absolutely
understandable
due that you are able to control Wifi BT AirplaneMode data etc.

This patch takes care on this problem and adds an option to control this
behaviour.

ATTENTION: Due that this is really an UX derp for an average user this
feature is
enabled by default now. User need to disable it explicit in settings to
get on secure
locksreen the original android behaviour back.

SystemUI: fix quick settings on secure lockscreen feature
by Altaf-Mahdi

With the current implementation there was a bug were if you expand a
notification
you will be able to access quick settings like normal.

A more simpler solution is to use the existing function
setQsExpansionEnabled() which works

Change-Id: I8ae2c2b540c9fe0fdc2fbd0f0e3583e4f8c1e29d
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>