base: don't show reboot options for safe mode
In a user long press in "power off" menu the system lead it to "reboot to safe mode", but if
the our advance power menu is enabled, we are showing the reboot options instead of the
safe mode advice message. Since reboot options doesn't made sense here is better to
respect the original behavior.

Change-Id: I45e7f7a780655ed79d442f583fd9ffa6fba4e447
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>