ListPopupWindow: Correct a negative height before showing
Symptom:
If an application set a negative height to the popup list,
surfaceflinger is crashed with SIGABRT.

Root cause:
WindowManagerService dose not expect negative
height of ListPopupWindow. If it's negative,
WindowManagerService set the negative value to GraphicBufferAlloc,
but GraphicBufferAlloc handle the value as unsigned int,
then surfaceflinger is crashed with SIGABRT.

Solution:
The popup is drawn with 1px height if mDropDownHeight is negative.

Bug: 33441454

Change-Id: I5887674d302e567abfe66147de4819cfdf0ef97b
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>