Fast paths for hwui matrix and gradient shaders
The shaders for matrix and simple gradients do many operations which
are not necessary for the most common values, and the matrix shader
is extremely shader heavy (causes noticeable slowdown on low-end GPUs
for e.g. the contacts application). This patch detects these common
inputs and uses much faster shaders instead. It also optimises the
un-premultiply and re-multiply for the general matrix shader, which
should in theory be as accurate (possibly more accurate).

Test: manual - visual inspection on fugu (nexus player)

Change-Id: I6b5fb4ee6f01aea37d09598d09ef86ee8135e39d
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>