Resize new created stack to fullscreen after docked stack tasks has moved into it.
The mFullscreen of new created stack will determined by WMS, and if the docked stack
is exists it will be false.And in the moveTasksToFullscreenStackLocked workflow the resize
stack action is early than the fullscreen stack create time, while move the tasks of docked
stack to the fullscreen stack, we need to create it first, at this time, docked stack is
exists the new created stack mFullscreen will be false. It will make the tasks launch in
this stack keep show in multi-window UI style, but its need to be fullscreen.

So, we need to do the resizeStack action after the move mission is done, ensure that the
new created stack will be resize to fullscreen.

Test: multi-window related functions works fine

Change-Id: I4eb706f4f391831360c79ee990793fb12b202017
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>